refactor: remove dead recording handler and extract two static helpers from App.xaml.cs (#493)
* refactor: remove dead OnRecordingStateChanged handler and its localization strings

NodeService.RecordingStateChanged was never subscribed in App.xaml.cs,
so the handler and its six Activity_Recording* resource keys were unreachable
at runtime. Removing them and the matching entries from all five locale
resw files so localization tests remain green.


* refactor: extract AppRunMarker from App.xaml.cs

The three static run-marker methods (CheckPreviousRun, MarkRunStarted,
MarkRunEnded) depended only on a file path. Moving them to a dedicated
class removes 37 lines from App and gives the behavior a named home.


* refactor: extract CliUninstallHandler from App.xaml.cs

The --uninstall CLI path (RunCliUninstallAsync, CliRedact, AttachConsole
P/Invoke) had no dependency on App instance state. Moving it to a
dedicated static class removes ~150 lines from App and gives the
headless uninstall entry point a clear, named home.


---------
