+++ src/OpenClaw.Tray.WinUI/App.xaml.cs
@@ -22,7 +22,6 @@ using System.Drawing;
 using System.IO;
 using System.IO.Pipes;
 using System.Linq;
-using System.Runtime.InteropServices;
 using System.Text;
 using System.Text.Json;
 using System.Threading;
@@ -279,7 +278,7 @@ public partial class App : Application, OpenClawTray.Services.IAppCommands
             ?? Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData),
         "OpenClawTray");
     private static readonly string CrashLogPath = Path.Combine(DataPath, "crash.log");
-    private static readonly string RunMarkerPath = Path.Combine(DataPath, "run.marker");
+    private static readonly AppRunMarker s_runMarker = new(Path.Combine(DataPath, "run.marker"));
 
     public App()
     {
@@ -297,8 +296,8 @@ public partial class App : Application, OpenClawTray.Services.IAppCommands
 
         InitializeComponent();
         
-        CheckPreviousRun();
-        MarkRunStarted();
+        s_runMarker.Check();
+        s_runMarker.MarkStarted();
         
         // Hook up crash handlers
         this.UnhandledException += OnUnhandledException;
@@ -326,7 +325,7 @@ public partial class App : Application, OpenClawTray.Services.IAppCommands
     
     private void OnProcessExit(object? sender, EventArgs e)
     {
-        MarkRunEnded();
+        s_runMarker.MarkEnded();
         try
         {
             Logger.Info($"Process exiting (ExitCode={Environment.ExitCode})");
@@ -359,192 +358,6 @@ public partial class App : Application, OpenClawTray.Services.IAppCommands
             }
         }
         catch { /* Ignore logging failures */ }
-    }
-
-    // -----------------------------------------------------------------------
-    // CLI uninstall path
-    // Invoked when --uninstall is present in argv. Runs headlessly without
-    // creating the tray UI. Attaches to the parent console so stdout/stderr
-    // are visible when invoked from PowerShell or cmd.
-    // -----------------------------------------------------------------------
-
-    [DllImport("kernel32.dll", SetLastError = true)]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    private static extern bool AttachConsole(int dwProcessId);
-
-    private const int AttachParentProcess = -1;
-
-    private static async Task RunCliUninstallAsync(string[] args)
-    {
-        // Attach to parent console so output is visible when invoked from
-        // PowerShell or cmd.  Fails silently if no parent console exists.
-        AttachConsole(AttachParentProcess);
-
-        bool dryRun            = args.Contains("--dry-run",            StringComparer.OrdinalIgnoreCase);
-        bool confirmDestructive = args.Contains("--confirm-destructive", StringComparer.OrdinalIgnoreCase);
-
-        // Locate --json-output <path> argument
-        string? jsonOutputPath = null;
-        for (int i = 0; i < args.Length - 1; i++)
-        {
-            if (string.Equals(args[i], "--json-output", StringComparison.OrdinalIgnoreCase))
-            {
-                jsonOutputPath = args[i + 1];
-                break;
-            }
-        }
-
-        if (!confirmDestructive && !dryRun)
-        {
-            Console.Error.WriteLine(
-                "ERROR: --uninstall requires --confirm-destructive (or --dry-run).");
-            Environment.Exit(2);
-            return;
-        }
-
-        var settings = new SettingsManager();
-        var engine   = LocalGatewayUninstall.Build(settings, logger: new AppLogger());
-
-        LocalGatewayUninstallResult result;
-        try
-        {
-            result = await engine.RunAsync(new LocalGatewayUninstallOptions
-            {
-                DryRun             = dryRun,
-                ConfirmDestructive = confirmDestructive
-            });
-        }
-        catch (Exception ex)
-        {
-            Console.Error.WriteLine($"ERROR: Uninstall engine threw: {ex.Message}");
-            Environment.Exit(1);
-            return;
-        }
-
-        // Human-readable summary (tokens already redacted inside engine steps)
-        Console.WriteLine("OpenClaw Local Gateway Uninstall");
-        Console.WriteLine($"DryRun:   {dryRun}");
-        Console.WriteLine($"Success:  {result.Success}");
-        Console.WriteLine($"Steps:    {result.Steps.Count} ({result.SkippedSteps.Count} skipped)");
-        Console.WriteLine($"Errors:   {result.Errors.Count}");
-        foreach (var e in result.Errors)
-            Console.Error.WriteLine($"  ERROR: {CliRedact(e)}");
-        Console.WriteLine("Postconditions:");
-        Console.WriteLine($"  WslDistroAbsent:    {result.Postconditions.WslDistroAbsent}");
-        Console.WriteLine($"  AutostartCleared:   {result.Postconditions.AutostartCleared}");
-        Console.WriteLine($"  SetupStateAbsent:   {result.Postconditions.SetupStateAbsent}");
-        Console.WriteLine($"  DeviceTokenCleared: {result.Postconditions.DeviceTokenCleared}");
-        Console.WriteLine($"  McpTokenPreserved:  {result.Postconditions.McpTokenPreserved}");
-        Console.WriteLine($"  KeepalivesAbsent:   {result.Postconditions.KeepalivesAbsent}");
-        Console.WriteLine($"  VhdDirAbsent:       {result.Postconditions.VhdDirAbsent}");
-        Console.WriteLine($"  LocalGatewayRecordsAbsent:      {result.Postconditions.LocalGatewayRecordsAbsent}");
-        Console.WriteLine($"  LocalGatewayIdentityDirsAbsent: {result.Postconditions.LocalGatewayIdentityDirsAbsent}");
-
-        // JSON output — redaction applied to step details and error strings
-        if (jsonOutputPath != null)
-        {
-            try
-            {
-                var dir = Path.GetDirectoryName(jsonOutputPath);
-                if (!string.IsNullOrEmpty(dir) && !Directory.Exists(dir))
-                    Directory.CreateDirectory(dir);
-
-                var payload = new
-                {
-                    success = result.Success,
-                    dry_run = dryRun,
-                    steps   = result.Steps.Select(s => new
-                    {
-                        name   = s.Name,
-                        status = s.Status.ToString(),
-                        detail = CliRedact(s.Detail)
-                    }),
-                    errors       = result.Errors.Select(CliRedact),
-                    skipped_steps = result.SkippedSteps,
-                    postconditions = new
-                    {
-                        wsl_distro_absent     = result.Postconditions.WslDistroAbsent,
-                        autostart_cleared     = result.Postconditions.AutostartCleared,
-                        setup_state_absent    = result.Postconditions.SetupStateAbsent,
-                        device_token_cleared  = result.Postconditions.DeviceTokenCleared,
-                        mcp_token_preserved   = result.Postconditions.McpTokenPreserved,
-                        keepalives_absent     = result.Postconditions.KeepalivesAbsent,
-                        vhd_dir_absent        = result.Postconditions.VhdDirAbsent,
-                        local_gateway_records_absent = result.Postconditions.LocalGatewayRecordsAbsent,
-                        local_gateway_identity_dirs_absent = result.Postconditions.LocalGatewayIdentityDirsAbsent
-                    }
-                };
-
-                File.WriteAllText(jsonOutputPath, JsonSerializer.Serialize(
-                    payload, new JsonSerializerOptions { WriteIndented = true }));
-
-                Console.WriteLine($"JSON result: {jsonOutputPath}");
-            }
-            catch (Exception ex)
-            {
-                Console.Error.WriteLine(
-                    $"WARNING: Failed to write JSON output to '{jsonOutputPath}': {ex.Message}");
-            }
-        }
-
-        Environment.Exit(result.Success ? 0 : 1);
-    }
-
-    /// <summary>
-    /// Redacts token/key material from a string before writing it to CLI
-    /// stdout or a JSON output file.  Mirrors the PowerShell Invoke-Redact
-    /// pattern in validate-wsl-gateway-uninstall.ps1.
-    /// </summary>
-    private static string? CliRedact(string? value)
-    {
-        if (string.IsNullOrEmpty(value)) return value;
-        // Redact JSON field values for known secret fields.
-        value = System.Text.RegularExpressions.Regex.Replace(
-            value,
-            @"(""(?i:deviceToken|device_token|token|bootstrapToken|bootstrap_token|PrivateKeyBase64|PublicKeyBase64)""\s*:\s*"")[^""]+("")",
-            "$1<redacted>$2");
-        // Redact bare key=value / key: value patterns.
-        value = System.Text.RegularExpressions.Regex.Replace(
-            value,
-            @"(?i)((?:device|bootstrap|gateway|auth|mcp)[_-]?token\s*[:=]\s*)[^\s,""'}{]+",
-            "$1<redacted>");
-        return value;
-    }
-    
-    private static void CheckPreviousRun()
-    {
-        try
-        {
-            if (File.Exists(RunMarkerPath))
-            {
-                var startedAt = File.ReadAllText(RunMarkerPath);
-                Logger.Error($"Previous session did not exit cleanly (started {startedAt})");
-                File.Delete(RunMarkerPath);
-            }
-        }
-        catch { }
-    }
-    
-    private static void MarkRunStarted()
-    {
-        try
-        {
-            var dir = Path.GetDirectoryName(RunMarkerPath);
-            if (!string.IsNullOrEmpty(dir) && !Directory.Exists(dir))
-                Directory.CreateDirectory(dir);
-            File.WriteAllText(RunMarkerPath, DateTime.Now.ToString("O"));
-        }
-        catch { }
-    }
-    
-    private static void MarkRunEnded()
-    {
-        try
-        {
-            if (File.Exists(RunMarkerPath))
-                File.Delete(RunMarkerPath);
-        }
-        catch { }
     }
 
     private void OnUiThread(Microsoft.UI.Dispatching.DispatcherQueueHandler action) => _dispatcherQueue?.TryEnqueue(action);
@@ -582,7 +395,7 @@ public partial class App : Application, OpenClawTray.Services.IAppCommands
         // -----------------------------------------------------------------------
         if (_startupArgs.Contains("--uninstall", StringComparer.OrdinalIgnoreCase))
         {
-            await RunCliUninstallAsync(_startupArgs);
+            await CliUninstallHandler.RunAsync(_startupArgs);
             return; // Environment.Exit called inside; defensive return
         }
 
@@ -2127,30 +1940,6 @@ public partial class App : Application, OpenClawTray.Services.IAppCommands
                     deviceId);
             }
             catch { /* ignore */ }
-        }
-    }
-    
-    private void OnRecordingStateChanged(object? sender, RecordingStateEventArgs args)
-    {
-        var source = args.Type == RecordingType.Screen ? "Screen" : "Camera";
-        if (args.IsActive)
-        {
-            var title = args.Type == RecordingType.Screen
-                ? LocalizationHelper.GetString("Activity_ScreenRecordingStarted")
-                : LocalizationHelper.GetString("Activity_CameraRecordingStarted");
-            var duration = args.DurationMs > 0 ? $" ({args.DurationMs / 1000.0:0.#}s)" : "";
-            AddRecentActivity($"{title}{duration}", category: "node",
-                icon: "🔴",
-                details: string.Format(LocalizationHelper.GetString("Activity_RecordingRequestedByAgent"), source));
-        }
-        else
-        {
-            var title = args.Type == RecordingType.Screen
-                ? LocalizationHelper.GetString("Activity_ScreenRecordingComplete")
-                : LocalizationHelper.GetString("Activity_CameraRecordingComplete");
-            AddRecentActivity(title, category: "node",
-                icon: "✅",
-                details: string.Format(LocalizationHelper.GetString("Activity_RecordingSentToAgent"), source));
         }
     }
